worked on image uploads some more
